+++ CHANGELOG.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-11-20
+### Added
+- Rename nodes for other langs + combine otp-ui locs.
+- Add json generator for other langs.
+- Rename source loc files in .sh script; process multiple localization files for each lang.
+
+### Fixed
+- Docs CHANGELOG.
+- Docs.
+- Run junk added to poetry.
+- Run junk added to poetry.
+
+### Changed
+- Update changelog.
+- Initial project create.
+
 ## [0.1.0] - 2022-11-18
 ### Added
 - Rename source loc files in .sh script; process multiple localization files for each lang.
@@ -17,5 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial project create.
 
-[Unreleased]: https://github.com/OpenTransitTools/yaml-json-tools/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/OpenTransitTools/yaml-json-tools/compare/0.2.0...HEAD
+[0.2.0]: https://github.com/OpenTransitTools/yaml-json-tools/compare/0.1.0...0.2.0
 [0.1.0]: https://github.com/OpenTransitTools/yaml-json-tools/releases/tag/0.1.0